requirements.txt
requests==2.31.0
python-dotenv==1.0.0
click==8.1.7
python-gitlab==4.4.0
pytz==2023.3 